feat(billing): wall-clock cadence, and a plan with its prices in one call
Two changes that ship together because the second is built on the shared cadence
module the first introduces.

Wall-clock cadence
- `minute` joins the price intervals. A cadence is UNIT x COUNT, so `minute` x 10
  bills every ten minutes and `month` x 3 bills quarterly; there is no separate
  `quarterly` or `ten_minutely`. Calendar cadences (day/week/month/year) keep
  pinning to a date at 02:00 Africa/Lagos, while a wall-clock cadence steps an
  exact interval from the anchor instant.
- Cadence math moves into `@nombaone/core-contracts/billing` as the one
  implementation: `periodsPerYear`, `toMonthlyKobo`, `intervalLabel`,
  `intervalShort`. The console and the API had already drifted apart here (a
  daily price was valued x30/month in one and x365/12 in the other); they now
  share a single source, and a unit test pins the interval list against the
  Postgres enum so the two cannot separate again.
- Migration 0022 adds the enum value.

Plans and prices: one intent, one call
- `POST /v1/plans` now takes an optional `prices` array and creates the plan and
  its price rows in ONE transaction. Either the whole catalog entry lands or none
  of it does, so a plan can no longer exist with no price and no way to bill it.
  Guards run before the transaction opens, and references are pre-minted because
  `INSERT ... VALUES (a),(b) RETURNING` does not guarantee row order and callers
  will zip the request and response arrays.
- Embedding prices requires `prices:write` in addition to `plans:write`. Without
  that check a `plans:write`-only key could mint price rows through the plan body.
- Additive: the nested `POST /v1/plans/{id}/prices` is unchanged, and a bare
  create behaves exactly as before (`data.prices` is then `[]`).
- Console: the surface is "Plans", not "Plans and prices" - the split was a data
  decision that had leaked into the merchant's mental model. One form now creates
  the plan and every cadence it sells on: set the monthly price and the annual one
  is derived from it, discount and all, as a suggestion the merchant can edit.
  Plans render a price ladder with the saving, and show real subscriber and MRR
  counts instead of hardcoded zeros.

Prices stay immutable. Raising one mints a new row and retires the old, so
existing subscribers keep the price they signed up on - the console now says so
plainly instead of hiding it as a mechanic.

+++ packages/core-contracts/src/validations/billing-settings.ts
@@ -5,13 +5,27 @@ import { z } from 'zod';
  * Every field optional (PUT patches only supplied keys). Cross-field refinement:
  * the max dunning window must be ≥ the largest configured interval, else attempts
  * exhaust before the schedule can play out.
+ *
+ * `dunningIntervalsHours` is FRACTIONAL (`0.25` is fifteen minutes). A retry ladder has
+ * to fit inside the billing period it is recovering — the renewal sweep skips `past_due`
+ * subscriptions, so a ladder longer than the cadence freezes renewals until dunning
+ * resolves. With a sub-day cadence (`interval: 'minute'`) an integer-hour floor made a
+ * fitting ladder impossible to express: the first decline on a ten-minute plan would
+ * freeze it for an hour at best, 24h by default. It is the ONLY fractional field here —
+ * it is backed by `jsonb`, while `dunning_max_window_hours` and `grace_period_hours` are
+ * `integer` columns, so accepting a fraction for those would only produce a value
+ * Postgres rejects. Neither needs to be sub-hour: exhaustion trips on
+ * `dunningMaxAttempts` long before a 1-hour window, and grace already accepts 0.
+ *
+ * Note that `paydayBiasEnabled` snaps a retry forward onto a calendar payday at 02:00,
+ * which swamps any sub-day ladder — turn it OFF on a tenant billing sub-daily.
  */
 export const updateBillingSettingsBody = z
   .object({
     partialCollectionEnabled: z.boolean().optional(),
     prorationCreditPolicy: z.enum(['credit_next_cycle', 'none']).optional(),
     dunningMaxAttempts: z.coerce.number().int().min(1).max(10).optional(),
-    dunningIntervalsHours: z.array(z.coerce.number().int().positive()).min(1).optional(),
+    dunningIntervalsHours: z.array(z.coerce.number().positive()).min(1).optional(),
     dunningMaxWindowHours: z.coerce.number().int().positive().optional(),
     gracePeriodHours: z.coerce.number().int().min(0).optional(),
     paydayDays: z.array(z.coerce.number().int().min(1).max(31)).min(1).optional(),
